refactor(apps/server): extract legacy payment admission under admission::legacy (D1)
First decoupling slice of the payment/policy excision plan: move the
cohesive payment-only clusters out of unified_server.rs into an
explicit module tree so the remaining surface can be deleted
independently of the PIR serving path. Pure relocation, no behavior
change; call sites keep working through narrow use statements.

Moved, verbatim:

- 134 lines of Cashu credential loading/validation helpers
  (load_cashu_epoch_keys_v1, parse_cashu_exposure_limits_v1,
  validate_existing_private_sqlite_path_v1) to admission/legacy/cashu.rs
- 1038 lines of strict admission assembly (ARC policy validation,
  load_strict_service_admission_v1, BAT V2 forbidden-config check,
  cashu runtime configuration validation) to
  admission/legacy/strict_admission.rs

Still inside unified_server.rs (next slices): StrictServiceAdmissionRuntimeV1
state + per-connection grant wiring, policy endpoint handlers, the
payment half of the CLI surface, and the service-gate opcode logic.

Verified: cargo check -p runtime --all-targets (default, cuckoo-oram,
standard-cashu-process-e2e, shared-issuer-process-e2e) and
cargo test -p runtime all green.
